refactor: group and exgtract similar app bars
