+++ src/components/src/AppBar/src/PopToTopAppBar.tsx
@@ -5,12 +5,7 @@ import {NativeStackHeaderProps} from "@react-navigation/native-stack";
 import {AppBar} from "./AppBar";
 import {AppBarButton} from "./AppBarButton";
 
-export default function RegisterAppBar({navigation}: NativeStackHeaderProps) {
-  const onCancel = () => {
-    const currentIndex = navigation?.getState().index;
-    navigation?.pop(currentIndex - 1);
-  };
-
+export default function PopToTopAppBar({navigation}: NativeStackHeaderProps) {
   return (
     <AppBar.Container>
       <AppBar.Header>
@@ -18,7 +13,7 @@ export default function RegisterAppBar({navigation}: NativeStackHeaderProps) {
           <AppBarButton.HistoryBack navigation={navigation} />
         </AppBar.Left>
         <AppBar.Right marginRight={24}>
-          <AppBarButton.Cancel onPress={onCancel} />
+          <AppBarButton.Cancel navigation={navigation} />
         </AppBar.Right>
       </AppBar.Header>
     </AppBar.Container>
