+++ src/navigation/stack/src/HomeStack.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import {createNativeStackNavigator} from "@react-navigation/native-stack";
 
-import {HomeAppBar, NotificationsAppBar} from "@/components";
+import {HomeAppBar, SimpleAppBar} from "@/components";
 import {stackScreenOptions} from "@/navigation/config";
 import {HomeTabParamList} from "@/navigation/navigation";
 import {HomeScreen, NotificationsScreen} from "@/screens/home";
@@ -16,7 +16,7 @@ export default function HomeStack() {
         <Stack.Screen name="HomeScreen" options={{header: HomeAppBar}} component={HomeScreen} />
         <Stack.Screen
           name="NotificationsScreen"
-          options={{header: NotificationsAppBar}}
+          options={{header: SimpleAppBar, title: "알림"}}
           component={NotificationsScreen}
         />
       </Stack.Group>
